Move pipeline functions into a separate package
